fix: Removed Contact form
Removed contact form in about.html becuase we were unable to get the
completion messages to display on the same html page without refreshing
the page. app.js was added and mailer.php was added for the near
future. PHP sends mail. App.js contains the AJAX and JQuery neccessary for
the contact form.