feat: Add working contact form
The working contact form has been added to the abouts.html. The contact
form uses AJAX and js found in app.js to send values from the contact form
to a php script found in mailer.php. The php script check if all data is
received and sends the email to info@sportomaticapp.com and displays the
corresponding succes or failure messages on the about.html page.